Add files via upload
This program experiments with CPU scheduling algorithms (SJF, RR, PR, and FIFO). It takes a file which contains data for which algorithm to use and data used for the CPU/IO bursts, and outputs the total cpu utilization, throughput, average turnaround time, and average waiting time in the ready queue for each process control block.